New madminer.analysis class with general data analysis features
