+++ madminer/sampling.py
@@ -5,6 +5,7 @@ import numpy as np
 import collections
 import six
 
+from madminer.analysis import DataAnalyzer
 from madminer.utils.interfaces.madminer_hdf5 import load_madminer_settings, madminer_event_loader
 from madminer.utils.interfaces.madminer_hdf5 import save_preformatted_events_to_madminer_file
 from madminer.utils.analysis import get_theta_value, get_theta_benchmark_matrix, get_dtheta_benchmark_matrix
@@ -16,7 +17,7 @@ from madminer.utils.various import format_benchmark, create_missing_folders, shu
 logger = logging.getLogger(__name__)
 
 
-class SampleAugmenter:
+class SampleAugmenter(DataAnalyzer):
     """
     Sampling and data augmentation.
 
@@ -74,85 +75,7 @@ class SampleAugmenter:
     """
 
     def __init__(self, filename, disable_morphing=False, include_nuisance_parameters=True):
-        # Save setup
-        self.include_nuisance_parameters = include_nuisance_parameters
-        self.madminer_filename = filename
-
-        logger.info("Loading data from %s", filename)
-
-        # Load data
-        (
-            self.parameters,
-            self.benchmarks,
-            self.benchmark_is_nuisance,
-            self.morphing_components,
-            self.morphing_matrix,
-            self.observables,
-            self.n_samples,
-            _,
-            self.reference_benchmark,
-            self.nuisance_parameters,
-        ) = load_madminer_settings(filename, include_nuisance_benchmarks=include_nuisance_parameters)
-
-        self.n_parameters = len(self.parameters)
-        self.n_benchmarks = len(self.benchmarks)
-        self.n_benchmarks_phys = np.sum(np.logical_not(self.benchmark_is_nuisance))
-
-        self.n_nuisance_parameters = 0
-        if self.nuisance_parameters is not None and include_nuisance_parameters:
-            self.n_nuisance_parameters = len(self.nuisance_parameters)
-        else:
-            self.nuisance_parameters = None
-
-        logger.info("Found %s parameters", self.n_parameters)
-        for key, values in six.iteritems(self.parameters):
-            logger.debug(
-                "   %s (LHA: %s %s, maximal power in squared ME: %s, range: %s)",
-                key,
-                values[0],
-                values[1],
-                values[2],
-                values[3],
-            )
-
-        if self.nuisance_parameters is not None:
-            logger.info("Found %s nuisance parameters", self.n_nuisance_parameters)
-            for key, values in six.iteritems(self.nuisance_parameters):
-                logger.debug("   %s (%s)", key, values)
-        else:
-            logger.info("Did not find nuisance parameters")
-
-        logger.info("Found %s benchmarks, of which %s physical", self.n_benchmarks, self.n_benchmarks_phys)
-        for (key, values), is_nuisance in zip(six.iteritems(self.benchmarks), self.benchmark_is_nuisance):
-            if is_nuisance:
-                logger.debug("   %s: nuisance parameter", key)
-            else:
-                logger.debug("   %s: %s", key, format_benchmark(values))
-
-        logger.info("Found %s observables", len(self.observables))
-        for i, obs in enumerate(self.observables):
-            logger.debug("  %2.2s %s", i, obs)
-        logger.info("Found %s events", self.n_samples)
-
-        # Morphing
-        self.morpher = None
-        if self.morphing_matrix is not None and self.morphing_components is not None and not disable_morphing:
-            self.morpher = PhysicsMorpher(self.parameters)
-            self.morpher.set_components(self.morphing_components)
-            self.morpher.set_basis(self.benchmarks, morphing_matrix=self.morphing_matrix)
-
-            logger.info("Found morphing setup with %s components", len(self.morphing_components))
-
-        else:
-            logger.info("Did not find morphing setup.")
-
-        # Nuisance morphing
-        self.nuisance_morpher = None
-        if self.nuisance_parameters is not None:
-            self.nuisance_morpher = NuisanceMorpher(
-                self.nuisance_parameters, list(self.benchmarks.keys()), self.reference_benchmark
-            )
-            logger.info("Found nuisance morphing setup")
+        super(SampleAugmenter, self).__init__(filename, disable_morphing, include_nuisance_parameters)
 
     def extract_samples_train_plain(
         self, theta, n_samples, folder=None, filename=None, test_split=0.2, switch_train_test_events=False
@@ -1008,59 +931,6 @@ class SampleAugmenter:
 
         return all_thetas, all_xsecs, all_xsec_uncertainties
 
-    def extract_raw_data(self, theta=None, derivative=False):
-
-        """
-        Returns all events together with the benchmark weights (if theta is None) or weights for a given theta.
-
-        Parameters
-        ----------
-        theta : None or ndarray or str, optional
-            If None, the function returns all benchmark weights. If str, the function returns the weights for a given
-            benchmark name. If ndarray, it uses morphing to calculate the weights for this value of theta. Default
-            value: None.
-
-        derivative : bool, optional
-            If True and if theta is not None, the derivative of the weights with respect to theta are returned. Default
-            value: False.
-
-        Returns
-        -------
-        x : ndarray
-            Observables with shape `(n_unweighted_samples, n_observables)`.
-
-        weights : ndarray
-            If theta is None and derivative is False, benchmark weights with shape
-            `(n_unweighted_samples, n_benchmarks_phys)` in pb. If theta is not None and derivative is True, the gradient of
-            the weight for the given parameter with respect to theta with shape `(n_unweighted_samples, n_gradients)`
-            in pb. Otherwise, weights for the given parameter theta with shape `(n_unweighted_samples,)` in pb.
-
-        """
-
-        x, weights_benchmarks = next(madminer_event_loader(self.madminer_filename, batch_size=None))
-
-        if theta is None:
-            return x, weights_benchmarks
-
-        elif isinstance(theta, six.string_types):
-            i_benchmark = list(self.benchmarks.keys()).index(theta)
-            return x, weights_benchmarks[:, i_benchmark]
-
-        elif derivative:
-            dtheta_matrix = get_dtheta_benchmark_matrix("morphing", theta, self.benchmarks, self.morpher)
-
-            gradients_theta = mdot(dtheta_matrix, weights_benchmarks)  # (n_gradients, n_samples)
-            gradients_theta = gradients_theta.T
-
-            return x, gradients_theta
-
-        else:
-            theta_matrix = get_theta_benchmark_matrix("morphing", theta, self.benchmarks, self.morpher)
-
-            weights_theta = mdot(theta_matrix, weights_benchmarks)
-
-            return x, weights_theta
-
     def _sample(
             self,
             sets,
@@ -1566,65 +1436,6 @@ class SampleAugmenter:
             logger.info("Effective number of samples: %s", all_effective_n_samples[0])
 
         return all_x, all_augmented_data, all_thetas
-
-    def _calculate_benchmark_xsecs(self, start_event, end_event, use_nuisance_parameters):
-        xsecs_benchmarks = 0.0
-        squared_weight_sum_benchmarks = 0.0
-        n_observables = 0
-        for obs, weights in madminer_event_loader(
-            self.madminer_filename,
-            start=start_event,
-            end=end_event,
-            include_nuisance_parameters=use_nuisance_parameters,
-            benchmark_is_nuisance=self.benchmark_is_nuisance,
-        ):
-            xsecs_benchmarks += np.sum(weights, axis=0)
-            squared_weight_sum_benchmarks += np.sum(weights * weights, axis=0)
-            n_observables = obs.shape[1]
-        return xsecs_benchmarks, squared_weight_sum_benchmarks, n_observables
-
-    def _train_test_split(self, train, test_split):
-        """
-        Returns the start and end event for train samples (train = True) or test samples (train = False).
-
-        Parameters
-        ----------
-        train : bool
-            True if training data is generated, False if test data is generated.
-
-        test_split : float
-            Fraction of events reserved for testing.
-
-        Returns
-        -------
-        start_event : int
-            Index of the first unweighted event to consider.
-
-        end_event : int
-            Index of the last unweighted event to consider.
-
-        """
-        if train:
-            start_event = 0
-
-            if test_split is None or test_split <= 0.0 or test_split >= 1.0:
-                end_event = None
-            else:
-                end_event = int(round((1.0 - test_split) * self.n_samples, 0))
-                if end_event < 0 or end_event > self.n_samples:
-                    raise ValueError("Irregular train / test split: sample {} / {}", end_event, self.n_samples)
-
-        else:
-            if test_split is None or test_split <= 0.0 or test_split >= 1.0:
-                start_event = 0
-            else:
-                start_event = int(round((1.0 - test_split) * self.n_samples, 0)) + 1
-                if start_event < 0 or start_event > self.n_samples:
-                    raise ValueError("Irregular train / test split: sample {} / {}", start_event, self.n_samples)
-
-            end_event = None
-
-        return start_event, end_event
 
 
 def combine_and_shuffle(input_filenames, output_filename, k_factors=None, overwrite_existing_file=True):
